Moved movie_detail_page folder into pages folder
