+++ src/components/pages/movie_detail_page/LinksToExternalPagesGroup.js
