+++ src/components/pages/movie_detail_page/LoadingEmptyScreen.js
